# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Strom API",
    "description": "REST and WebSocket API for the Strom pipeline engine",
    "contact": {
      "name": "Per Enstedt",
      "email": "per.enstedt@eyevinn.se"
    },
    "license": {
      "name": "MIT OR Apache-2.0"
    },
    "version": "0.4.9"
  },
  "paths": {
...
    "/api/log-level": {
      "get": {
        "tags": [
          "System"
        ],
        "summary": "Get the current log level filter",
        "operationId": "get_log_level",
        "responses": {
          "200": {
            "description": "Current log level",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LogLevelResponse"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "System"
        ],
        "summary": "Set the log level filter at runtime",
        "operationId": "set_log_level",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetLogLevelRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Log level updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LogLevelResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid filter string",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ErrorResponse": {
        "type": "object",
        "description": "Standard error response.",
        "required": [
          "error"
        ],
        "properties": {
          "details": {
            "type": [
              "string",
              "null"
            ]
          },
          "error": {
            "type": "string"
          }
        }
      },
...
      "LogLevelResponse": {
        "type": "object",
        "description": "Response for log level queries and updates.",
        "required": [
          "current",
          "default"
        ],
        "properties": {
          "current": {
            "type": "string",
            "description": "The currently active log filter string (e.g. \"info,strom::api=debug\")"
          },
          "default": {
            "type": "string",
            "description": "The default filter the server started with"
          }
        }
      },
...
      "SetLogLevelRequest": {
        "type": "object",
        "description": "Request to change the log level at runtime.",
        "required": [
          "filter"
        ],
        "properties": {
          "filter": {
            "type": "string",
            "description": "The new log filter string (e.g. \"info,strom::api=debug\")"
          }
        }
      },
...
    }
  },
...
}